+++ fuzzers/fuzz_anything/libafl_atheris/src/lib.rs
@@ -216,14 +216,9 @@ pub extern "C" fn LLVMFuzzerRunDriver(
             ExitKind::Ok
         };
 
+        let mut executor = ShadowExecutor::new(executor, tuple_list!(cmplog_observer));
         // Setup a tracing stage in which we log comparisons
-        let tracing = TracingStage::new(InProcessExecutor::new(
-            &mut harness,
-            tuple_list!(cmplog_observer),
-            &mut fuzzer,
-            &mut state,
-            &mut mgr,
-        )?);
+        let tracing = ShadowTracingStage::new();
 
         // Setup a randomic Input2State stage
         let i2s =
